backends/requirements.txt
aerich==0.7.1
aiofiles==22.1.0
asyncpg==0.26.0
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
easyocr==1.6.2
faiss-cpu==1.7.2
numpy==1.23.3
openai==0.23.0
pdf2image==1.16.0
python-dotenv==0.21.0
requests==2.28.1
sanic==22.6.2
Sanic_Cors==2.1.0
sentence-transformers==2.2.2
torch==1.12.1
tortoise-orm==0.19.2
